requirements.txt
# local package
-e .

# external requirements
alabaster==0.7.12
appdirs==1.4.3
appnope==0.1.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.28
Babel==2.6.0
backcall==0.1.0
black==18.6b4
bleach==2.1.4
boto3==1.9.16
botocore==1.12.18
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
coverage==4.5.1
cycler==0.10.0
dask==0.19.2
databricks-cli==0.8.2
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.2.3
feather-format==0.4.0
fire==0.1.3
flake8==3.5.0
Flask==1.0.2
gitdb2==2.0.4
GitPython==2.1.11
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
imagesize==1.1.0
imbalanced-learn==0.4.1
imblearn==0.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.3
mleap==0.8.1
mlflow==0.7.0
more-itertools==4.3.0
mypy==0.630
mypy-extensions==0.4.1
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
nose-exclude==0.5.0
notebook==5.7.0
numpy==1.15.2
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.8.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
py==1.7.0
pyarrow==0.11.0
pyasn1==0.4.4
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.1
pytest==3.9.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
querystring-parser==1.2.3
requests==2.20.0
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
smmap2==2.0.4
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
tabulate==0.8.2
terminado==0.8.1
testpath==0.3.1
toml==0.9.6
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xgboost==0.80
yapf==0.24.0